First pass, will add more if this is accepted.
This is my first pass at providing technical documention (examples) for the technical interfaces for blender.
It makes most sense to do this as part of the awesome repo (in my opinion).  Please let me know your thoughts about creating a branch on the master repo for programming language examples and allowing me to provide documentation to show others how to do the same & curating that branch of this repo.
Let me know your thoughts.